Use compensation image in tiled_image
Also moved some more code into the WSExperiment base-class
